requirements.txt
# From pypi
diraccfg
autopep8==1.3.3
coverage
requests>=2.9.1
mock>=1.0.1
pylint>=1.6.5
pytest>=3.6
pytest-cov>=2.2.0
pytest-mock
tornado>=5.0.0
